move the add_missing_data function to the cloud
